+++ app/templates/main/record-html-view.html
@@ -1,0 +1,5 @@
+<!-- Record HTML View -->
+<div class="govuk-tabs__panel govuk-tabs__panel--hidden govuk-tabs__panel__record-html-view"
+     id="record-html-view">
+    <h2 class="govuk-heading-l">Record HTML View</h2>
+</div>
