+++ app/templates/main/record-view.html
@@ -1,0 +1,33 @@
+{% if file_type == "iiif" %}
+    <div class="govuk-tabs__panel" id="record-view">
+        <div class="record-view-header">
+            <a href="{{ url_for('main.download_record', record_id=record['file_id']) }}"
+               class="govuk-button govuk-button__download--record record-view-download-btn"
+               data-module="govuk-button">Download record</a>
+        </div>
+        <div class="universal-viewer" id="viewer">
+            <script nonce="{{ csp_nonce() }}"
+                    src="https://cdn.jsdelivr.net/npm/universalviewer@4.0.25/dist/umd/UV.min.js"></script>
+            <link href="{{ url_for('static', filename='uv.min.css') }}"
+                  rel="stylesheet">
+            <noscript>
+                {% if file_extension == "pdf" %}
+                    <object data="{{ url_for('static', filename='files/temp_file.pdf') }}"
+                            width="100%"
+                            height="700px">
+                        <p>Please enable JS to view this record.</p>
+                    </object>
+                {% elif file_extension == "png" or file_extension == "jpg" %}
+                    <object data="{{ url_for('static', filename='files/temp_file.png') }}"
+                            width="100%"
+                            height="700px">
+                        <p>Please enable JS to view this record.</p>
+                    </object>
+                {% endif %}
+            </noscript>
+            <div id="uv" class="uv"></div>
+            <script id="init-uv" nonce="{{ csp_nonce() }}" manifest_url="{{ manifest_url }}" type="text/javascript" src="{{ url_for('static', filename='init.uv.js') }}">
+            </script>
+        </div>
+    </div>
+{% endif %}
