+++ app/templates/main/record.html
@@ -77,93 +77,11 @@
                             <li class="govuk-tabs__list-item">
                                 <a class="govuk-tabs__tab" href="#record-details">Record Details</a>
                             </li>
-                            <li class="govuk-tabs__list-item">
-                                <a class="govuk-tabs__tab" href="#record-html-view">Record HTML View</a>
-                            </li>
                         </ul>
                         <!-- Record View -->
-                        {% if file_type == "iiif" %}
-                            <div class="govuk-tabs__panel" id="record-view">
-                                <div class="record-view-header">
-                                    <a href="{{ url_for('main.download_record', record_id=record['file_id']) }}"
-                                       class="govuk-button govuk-button__download--record record-view-download-btn"
-                                       data-module="govuk-button">Download record</a>
-                                </div>
-                                <div class="universal-viewer" id="viewer">
-                                    <script nonce="{{ csp_nonce() }}"
-                                            src="https://cdn.jsdelivr.net/npm/universalviewer@4.0.25/dist/umd/UV.min.js"></script>
-                                    <link
-                                          href="{{ url_for('static', filename='uv.min.css') }}"
-                                          rel="stylesheet">
-                                    <noscript>
-                                        {% if file_extension == "pdf" %}
-                                            <object data="{{ url_for('static', filename='files/temp_file.pdf') }}"
-                                                    width="100%"
-                                                    height="700px">
-                                                <p></p>
-                                            </object>
-                                        {% elif file_extension == "png" or file_extension == "jpg" %}
-                                            <object data="{{ url_for('static', filename='files/temp_file.png') }}"
-                                                    width="100%"
-                                                    height="700px">
-                                                <p></p>
-                                            </object>
-                                        {% endif %}
-                                    </noscript>
-                                    <div id="uv" class="uv"></div>
-                                    <script id="init-uv" nonce="{{ csp_nonce() }}" manifest_url="{{ manifest_url }}" type="text/javascript" src="{{ url_for('static', filename='init.uv.js') }}">
-                                    </script>
-                                </div>
-                            </div>
-                        {% endif %}
+                        {% include "record-view.html" %}
                         <!-- Record Details -->
-                        <div class="govuk-tabs__panel govuk-tabs__panel--hidden govuk-tabs__panel__record-details"
-                             id="record-details">
-                            <div class="govuk-grid-column-two-thirds">
-                                <div class="details-container">
-                                    <h1 class="govuk-heading-m govuk-heading-m__record-details">Record details</h1>
-                                    <dl class="govuk-summary-list govuk-summary-list--record">
-                                        {% with dict = list_items %}
-                                            {% include "summary_list_items.html" %}
-                                        {% endwith %}
-                                    </dl>
-                                </div>
-                            </div>
-                            <div class="govuk-grid-column-one-third govuk-grid-column-one-third--record mobile-record-layout-rights">
-                                <div class="rights-container">
-                                    <h3 class="govuk-heading-m govuk-heading-m__rights-header">Rights to access</h3>
-                                    <a href="{{ url_for('main.download_record', record_id=record['file_id']) }}"
-                                       class="govuk-button govuk-button__download--record"
-                                       data-module="govuk-button">Download
-                                    record</a>
-                                    {% if download_filename %}
-                                        <p class="govuk-body govuk-body--download-filename">
-                                            The downloaded record will be named
-                                            <br>
-                                            <strong>{{ download_filename }}</strong>
-                                        </p>
-                                    {% endif %}
-                                    <p class="govuk-body govuk-body--terms-of-use">
-                                        Refer to <a href="/terms-of-use" class="govuk-link govuk-link--ayr">Terms of use.</a>
-                                    </p>
-                                </div>
-                            </div>
-                            <div class="govuk-grid-column-one-third govuk-grid-column-one-third--record mobile-record-layout-record-arrangement">
-                                <div class="record-container">
-                                    <h2 class="govuk-heading-m govuk-heading-m__record-header">Record arrangement</h2>
-                                    <ol class="record-arrangement-list">
-                                        {% for file_level in record["file_path"].split("/") %}
-                                            <li class="govuk-body__record-arrangement-list">{{ file_level }}</li>
-                                        {% endfor %}
-                                    </ol>
-                                </div>
-                            </div>
-                        </div>
-                        <!-- Record HTML View -->
-                        <div class="govuk-tabs__panel govuk-tabs__panel--hidden govuk-tabs__panel__record-html-view"
-                             id="record-html-view">
-                            <h2 class="govuk-heading-l">Record HTML View</h2>
-                        </div>
+                        {% include "record-details.html" %}
                     </div>
                 </div>
             </div>
